Adding update & delete functionality

+++ controller/userController.js
@@ -59,20 +59,54 @@ const userController = {
 
     //TODO PUT/DELETE endpoins
 
-    updateUser: async (request , response) => {
-        const serviceResponse = await userService.updateUser(id , data);
-        response.writeHead(200 , {"Content-Type" : "application/json"});
-        response.write("Hello from updateUser\n");
-        //TODO: extract the id and data from request body
-        return response.end();
+    updateUser: async (request , response , params) => {
+        try {
+
+            const id = params.id ;
+            const data = await fetchPostBodyData(request);
+
+            if (data.name === undefined && data.email === undefined) 
+                return Response(response , {status : 400 , data: {message : "No messages to update"}});
+            
+            //Check if user exist.
+            const user = await userService.getUserById(id);
+            if (user.length === 0) 
+                return Response(response , {status: 404 , data: {message: "User not found"}});
+            
+            //Check if email is already exist.
+            const isEmailExist = await userService.getEmail(data.email);
+            if (isEmailExist > 0) 
+                return Response(response , {status: 400 , data: {message: "Email is already exists"}});
+
+            const serviceResponse = await userService.updateUser(id , data);
+
+            return Response(response , {status: 200 , data: {message: "User updated sucessfully" , data: serviceResponse} });
+
+        } catch(error) {
+            console.log(error) ;
+            return Response(response , {status: 404 , data: {message: error.message}});
+        }
     },
 
-    deleteUser: async (request , response) => {
-        const serviceResponse = await userService.deleteUser(id);
-        response.writeHead(200 , {"Content-Type" : "application/json"});
-        response.write("Hello from deleteUser\n");
-        //TODO: extract the id from body request
-        return response.end();
+    deleteUser: async (request , response , params) => {
+        try {
+            const id = params.id;
+
+            const user = await userService.getUserById(id);
+            if (user.length === 0) 
+                return Response(response , {status: 404 , data: {message: "User not found"}});
+        
+            const serviceResponse = await userService.deleteUser(id);
+            
+            if (serviceResponse === 1)
+                return Response(response , {status: 200 , data: {message: "User deleted successfully"}});
+            else 
+                return Response(response , {status: 400 , data: {message: "User doesnot deleted , try again"}});
+
+        } catch(error) {
+            console.log(error) ;
+            return Response(response , {status: 404 , data: {message: error.message}});
+        }
     }
 }
